CalendarEntry: Use zod `.date()` for input `datetime-local`
This allows us to use the same validation schema on the server as well as on the frontend for client side validations.
The types form prisma and zod match.
However, there is special treatment for the way the `datetime-local` input works; which is at least contained in the LabeledTextField Component.