Fix Firefox datetime-local bug by splitting startAt in startDate and startTime
We use the input type date and time as separate inputs, which does work in Firefox.

We need to modify the zod Schema to ignore the startAt but handle the startDate and startTime.

We need to re-merge the separate info before saving. And we validate this value just in case.

The error component needed an update because this way to throw a ZodError results in a different object than the other cases.

To edit, we need to split the existing data again. Which is just way to complicated in JS!