Take 4b (useQuery return type)
According to https://tkdodo.eu/blog/react-query-and-type-script#the-four-generics the type signature for useQuery has 4 params, the third is the data.

BUT, it looks like our blitz useQuery has a completely different signature that tanstack useQuery. (Question asked in Discord)

We can still pass the return type and force the right response like this.