Fix logo source for logged in partner logos
This used the staging fall back all the time, because `APP_ORIGIN` was not present in react client side.

This duplicates `NEXT_PUBLIC_APP_ORIGIN` and uses that value.

The `isProduction` check was broken as well for the same reason, probably.

A follow up is to cleanup the likely unused `APP_ORIGIN`.