Work around eslint mess with package versions
Form some reason, we now need a very specific zod version to work with the eslint-plugin-react-compiler plugin.

But also, some Blitz features now show as errors in eslint which we need to work around.

And to make this less complex, the compiler is now on "warn" until we updated all of this…